Utilize attr_args meta for parsing
