+++ crates/ink/ir/src/ast/mod.rs
@@ -25,9 +25,13 @@
 //! the `value` part.
 
 mod attr_args;
+mod meta;
 
-pub use self::attr_args::{
-    AttributeArgs,
-    MetaNameValue,
-    PathOrLit,
+pub use self::{
+    attr_args::AttributeArgs,
+    meta::{
+        Meta,
+        MetaNameValue,
+        PathOrLit,
+    },
 };
